Extracted parts of lookup.js rendering into individual files; began work on dashboard.
