+++ jade_templates/includes/dashboard.jade
@@ -1,0 +1,8 @@
+h1 Wikiedit Monitor Dashboard
+div#main_graph_area
+  div#socket_graph
+  div#edit_graph
+  div#error_graph
+
+script
+   include dashboard.js
